pyproject.toml
[tool.poetry]
name = "telegram-asr-bot"
version = "0.1.0"
description = ""
authors = ["FrozenYogurtPuff <frozenyogurtpuff@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.11"
google-cloud-speech = "^2.13.1"
environs = "^9.5.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
